requirements.txt
pillow
tkinter
customtkinter